♻️ Feature: data tracker
* ♻️ Update interface and implementation for experiment tracking (#706)
Interface's methods are public. New methods in subclasses are private and used within the public methods. For not using methods in use cases that are defined in the infrastrucure.

* ♻️ Update use case for dataset versioning (#706)
When logging information of versioning for not use methods in use cases that are defined in infrastructure. Update unit tests

* ♻️ Update use case for fitting the data preprocessor (#706)
When logging preprocessor and the preprocessing steps to experiment tracking system. For not use methods in use cases that are defined in infrastructure. Update unit tests

* ♻️ Update use case for training the model (#706)
When logging model and hyperparameters to experiment tracking system. For not use methods in use cases that are defined in infrastructure.

* ♻️ Update use case for model evaluation (#706)
When loading model and logging metrics to experiment tracking system. For not use methods in use cases that are defined in infrastructure.

* ♻️ Update use case for model validation (#706)
When getting model metrics from experiment tracking system. For not use methods in use cases that are defined in infrastructure

* ♻️ Update use case for serving the model (#706)
When getting model path in storage from experiment tracking system. For not use methods in use cases that are defined in infrastructure.

* ♻️ Update data tracker and use cases (#706)
For not use implementation constants in use cases

PR
https://github.com/victorviro/ml_pipeline/pull/4